requirements.txt
Flask==1.1.2
mysqlclient==2.0.1
marshmallow==3.8.0
Flask-SQLALchemy==2.4.4
Pyjwt==1.4.2
pytest==6.1.1
flask-hashing==1.1
pytest-cov==2.10.1
Werkzeug==1.0.1
mysql-connector-python==8.0.22
StringGenerator==0.3.4
flask-cors==3.0.9